delete a package and move one component to other package
